(refactor) entity의 정보를 갖는 columnSpec 분리
